refactor: pokemon info
