+++ pokemon_app/lib/features/pokemon_info/view/pokemon_info_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:pokemon_app/repositories/pokemons_rep_export.dart';
-
+import '../widgets/widgets.dart';
 import '../bloc/pokemons_info_bloc.dart';
 
 class PokemonInfo extends StatefulWidget {
@@ -14,8 +14,6 @@ class PokemonInfo extends StatefulWidget {
 }
 
 class _InfoState extends State<PokemonInfo> {
-  PokemonInfoModel? pokemonInfo;
-
   final _pokemonsInfoBloc =
       PokemonsInfoBloc(GetIt.I<AbstractPokemonsInfoRep>());
 
@@ -39,7 +37,7 @@ class _InfoState extends State<PokemonInfo> {
           bloc: _pokemonsInfoBloc,
           builder: (context, state) {
             if (state is PokemonsInfoLoaded) {
-              return bodyWidget(context, state);
+              return PokemonInfoContent(state: state);
             }
             if (state is PokemonsInfoFailure) {
               return Center(
@@ -52,55 +50,9 @@ class _InfoState extends State<PokemonInfo> {
 
             return loaderWidget(context);
           },
-        )
-        );
+        ));
   }
 
-  loaderWidget(BuildContext context) =>
+  Widget loaderWidget(BuildContext context) =>
       const Center(child: CircularProgressIndicator());
-
-  bodyWidget(BuildContext context, state) => Stack(
-        children: <Widget>[
-          Positioned(
-            height: MediaQuery.of(context).size.height / 1.5,
-            width: MediaQuery.of(context).size.width - 30,
-            left: 15.0,
-            top: MediaQuery.of(context).size.height * 0.1,
-            child: Card(
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(15.0),
-              ),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: <Widget>[
-                  const SizedBox(),
-                  Hero(
-                      tag: state.pokemonsInfoLoaded.img,
-                      child: Container(
-                        height: 200.0,
-                        width: 200.0,
-                        decoration: BoxDecoration(
-                            image: DecorationImage(
-                                fit: BoxFit.cover,
-                                image: NetworkImage(
-                                    state.pokemonsInfoLoaded.img))),
-                      )),
-                  Text(
-                    state.pokemonsInfoLoaded.name,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                  ),
-                  Text("Height: ${state.pokemonsInfoLoaded.height}",
-                      style: Theme.of(context).textTheme.bodySmall),
-                  Text("Weight: ${state.pokemonsInfoLoaded.weight}",
-                      style: Theme.of(context).textTheme.bodySmall),
-                  Text(
-                    "Types: ${state.pokemonsInfoLoaded.types.join(', ')}",
-                    style: Theme.of(context).textTheme.bodySmall,
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      );
 }
