+++ pokemon_app/lib/features/pokemon_info/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'pokemon_info_content.dart';
